+++ src/main/java/com/ouaskanas/educonnect/Dao/Repositories/PostRepository.java
@@ -7,7 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface PostRepository extends JpaRepository<Post, Long> {
+public interface PostRepository extends JpaRepository<Post, Integer> {
     public Post findByTitle(String title);
-
 }
