+++ src/main/java/com/ouaskanas/educonnect/Dto/PostDto.java
@@ -14,7 +14,7 @@ import java.util.Date;
 public class PostDto {
     private String title;
     private String content;
-    private long author_id;
-    private long classroom_id;
+    private int author_id;
+    private int classroom_id;
     private boolean shared;
 }
