+++ src/main/java/com/ouaskanas/educonnect/Service/Manager/CommentManager.java
@@ -5,8 +5,8 @@ import com.ouaskanas.educonnect.Dao.Entities.Comment;
 import java.util.List;
 
 public interface CommentManager {
-    public List<Comment> getCommentFromPost(long post_id);
-    public Comment postComment(long post_id, Comment comment);
-    public boolean deleteComment(long comment_id, long post_id);
+    public List<Comment> getCommentFromPost(int post_id);
+    public Comment postComment(int post_id, Comment comment);
+    public boolean deleteComment(int comment_id, int post_id);
 
 }
